refactor: centralize vault mock data
