+++ src/types/vault.ts
@@ -1,0 +1,10 @@
+export interface VaultItem {
+  id: string
+  name: string
+  username: string
+  password: string
+  url?: string
+  folder?: string
+  notes?: string
+  updatedAt: string
+}
